RavenDB-21416 Move data archival patching tests to the other file DataArchivalPatchingTests.cs
